refactor(ui): extract TurnStart/TurnEnd arms to run_handlers::turn (dirge-4y4l stage 4)
The two plugin-gated turn arms (per-turn batcher reset/flush + on-turn-start/
on-turn-end hooks) moved to run_handlers/turn.rs (cfg(plugin) module),
deduping the tool-hook-slot clear into a helper. Completes the AgentEvent
arm extraction for run_interactive. ui/mod.rs 3770->3712. Behavior-
preserving; default + no-plugin both -D warnings clean, 2374 tests green.
